Fixed date format in threat-report-editor. Renamed report-editor folder threat-report-editor, since that folder also has a component named report-editor.
