+++ src/app/threat-dashboard/threat-dashboard.module.ts
@@ -36,10 +36,10 @@ import { ThreatReportNavigateGuard } from './threat-report-navigate.guard';
 import { ThreatReportOverviewService } from './services/threat-report-overview.service';
 import { ThreatReportSharedService } from './services/threat-report-shared.service';
 import { ReportTranslationService } from './services/report-translation.service';
-import { ThreatReportEditorComponent } from './report-editor/threat-report-editor.component';
-import { ReportEditorComponent } from './report-editor/report-editor/report-editor.component';
-import { ReportImporterComponent } from './report-editor/report-importer/report-importer.component';
-import { ReportUploadService } from './report-editor/report-importer/report-upload.service';
+import { ThreatReportEditorComponent } from './threat-report-editor/threat-report-editor.component';
+import { ReportEditorComponent } from './threat-report-editor/report-editor/report-editor.component';
+import { ReportImporterComponent } from './threat-report-editor/report-importer/report-importer.component';
+import { ReportUploadService } from './threat-report-editor/report-importer/report-upload.service';
 import { CollapsibleTreeComponent } from './collapsible-tree/collapsible-tree.component';
 import { KillChainTableComponent } from './kill-chain-table/kill-chain-table.component';
 import { RadarChartComponent } from './radar-chart/radar-chart.component';
