+++ src/app/threat-dashboard/threat-report-editor/file-upload/file-upload.component.scss
