+++ src/app/threat-dashboard/threat-report-editor/file-upload/upload.service.ts
