+++ src/app/threat-dashboard/threat-report-editor/report-editor/report-editor.component.scss
