+++ src/app/threat-dashboard/threat-report-editor/threat-report-editor.component.scss
